d_stage fixup

+++ include/d/d_com_inf_game.h
@@ -822,7 +822,7 @@ inline s8 dComIfGp_getStartStageLayer() {
     return g_dComIfG_gameInfo.play.getStartStageLayer();
 }
 
-inline s16 dComIfGp_getStartStagePoint() {
+inline s32 dComIfGp_getStartStagePoint() {
     return g_dComIfG_gameInfo.play.getStartStagePoint();
 }
